third_party/WebKit/LayoutTests/inspector/sources/debugger-pause/pause-in-internal-script.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>

function testFunction()
{
    var array = [2, 5, 7];
    var sum = 0;
    array.forEach(function(key)
    {
        sum += array[key];
    });
    return sum;
}

function test()
{
    SourcesTestRunner.runDebuggerTestSuite([
        function testSetBreakpoint(next)
        {
            SourcesTestRunner.showScriptSource("pause-in-internal-script.html", didShowScriptSource);

            var breakpointFunctionFrame = null;

            function didShowScriptSource(sourceFrame)
            {
                breakpointFunctionFrame = sourceFrame;
                TestRunner.addResult("Script source was shown.");
                SourcesTestRunner.setBreakpoint(sourceFrame, 13, "", true);
                SourcesTestRunner.runTestFunctionAndWaitUntilPaused(didPause);
            }

            function didPause(callFrames)
            {
                SourcesTestRunner.captureStackTrace(callFrames);
                SourcesTestRunner.removeBreakpoint(breakpointFunctionFrame, 13);
                next();
            }
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests that internal scripts unknown to front-end are processed correctly when appear in debugger call frames.
<a href="https://bugs.webkit.org/show_bug.cgi?id=64995">Bug 64995</a>
</p>

</body>
</html>
